+++ src/redturtle/volto/upgrades.py
@@ -9,6 +9,12 @@ from redturtle.volto.setuphandlers import remove_custom_googlebot
 from uuid import uuid4
 from zope.schema import getFields
 
+
+try:
+    from plone.base.utils import get_installer
+except Exception:
+    from Products.CMFPlone.utils import get_installer
+
 import json
 import logging
 
@@ -494,3 +500,11 @@ def to_4303(context):
     for brain in brains:
         event = brain.getObject()
         event.reindexObject(idxs=["start", "end"])
+
+
+def to_4305(context):
+    portal = api.portal.get()
+    installer = get_installer(portal, portal.REQUEST)
+    installer.uninstall_product(product_id="collective.volto.cookieconsent")
+    if not installer.is_product_installed("collective.volto.gdprcookie"):
+        installer.install_(product_id="collective.volto.gdprcookie")